fix ecommerce lineitem tracking in GA. smaller DOM footprint
Big changes to all the var and functions pre-init.  Most of these are
now in a single object, decreasing DOM footprint. Will likely pass this
into the control object and have it extended, so control functions
overwrite any placeholder functions (such as error handling) and the
initial object is extended to be used. this means myControl is likely to
get renamed.