change myControl to app
Lots of changes to variable names to make for shorter variables.
myControl changed to app. util changed to u and actions changed to a. So
myControl.action.showContent is now app.a.showContent

Initial variables (in loadscript and init.js) changed to app which is
then extended when class is instantiated.

Added support for META uri parameter. will update session with value