passive Q now auto-dispatches every 5 seconds
The passive Q will now auto-dispatch. This allows the developer to toss
non-essential requests into that Q that'll be dispatched leisurely.
Adding without executing a dispatchThis means the Q will not fire if an
immutable dispatch is in process.

Also, any category links that use the bindByAnchor/bindNav feature will
get an appCategoryDetail [max] request added to the passive Q. This'll
improve load time on those pages.